Add files via upload
The objective of this undertaking is to choose a suitable dataset following specified instructions, implement the data cleaning methods and techniques taught in the course, and subsequently refine and organize the chosen dataset to make it ready for essential analytical processes. In this project, a dataset related to Netflix userbase was employed for the designated tasks.